% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_decay_models.R
\name{add_critical_time}
\alias{add_critical_time}
\title{Add critical time to model parameters from biexponential fit}
\usage{
add_critical_time(parameters)
}
\arguments{
\item{parameters}{A data frame containing columns \code{p}, \code{r}, and \code{q}, 
typically returned by \code{fit_biexponential_model()}.}
}
\value{
A data frame with a new column:
\describe{
  \item{critical_time}{The estimated time \eqn{t_c} at which cultural memory begins to dominate.}
}
}
\description{
Computes the critical time \eqn{t_c} from the parameters \eqn{p}, \eqn{r}, and \eqn{q}
obtained via the biexponential model. This is the point in time at which the contributions
of communicative and cultural memory become equal.
}
\details{
The formula used is:
\deqn{
t_c = \frac{1}{p + r - q} \log \left( \frac{(p + r)(p - q)}{rq} \right)
}

This function is designed to work with the `params` output of \code{\link{fit_biexponential_model}}.
}
\seealso{
\code{\link{fit_biexponential_model}}
}
